+++ aioxmpp/tracking.py
@@ -30,6 +30,10 @@ actual tracking is not implemented here.
 
    This module was added in version 0.5.
 
+.. versionchanged:: 0.9
+
+   This module was completely rewritten in 0.9.
+
 .. seealso::
 
    Method :meth:`~.muc.Room.send_tracked_message`
@@ -63,30 +67,18 @@ class MessageState(Enum):
 
        This is a final state.
 
-    .. attribute:: TIMED_OUT
-
-       The tracking has timed out. Whether a timeout exists and how it is
-       handled depends on the tracking implementation.
-
-       This is a final state.
-
-    .. attribute:: CLOSED
-
-       The tracking itself got aborted and cannot make a statement about the
-       delivery of the stanza.
-
-       This is a final state.
-
     .. attribute:: ERROR
 
        An error reply stanza has been received for the stanza which was sent.
 
-       This is a final state.
+       This is, in most cases, a final state.
 
     .. attribute:: IN_TRANSIT
 
        The message is still queued for sending or has been sent to the peer
        server without stream management.
+
+       Depending on the tracking implementation, this may be a final state.
 
     .. attribute:: DELIVERED_TO_SERVER
 
@@ -97,8 +89,9 @@ class MessageState(Enum):
 
     .. attribute:: DELIVERED_TO_RECIPIENT
 
-       The message has been delivered to the recipient. Depending on the
-       tracking implementation, this may be a final state.
+       The message has been delivered to the recipient.
+
+       Depending on the tracking implementation, this may be a final state.
 
     .. attribute:: SEEN_BY_RECIPIENT
 
@@ -107,73 +100,174 @@ class MessageState(Enum):
 
     """
 
-    def __lt__(self, other):
-        if     ((other == MessageState.ABORTED or
-                 other == MessageState.CLOSED or
-                 other == MessageState.ERROR) and
-                self != MessageState.IN_TRANSIT):
-            return True
-        if     (other == MessageState.TIMED_OUT and
-                self != MessageState.IN_TRANSIT and
-                self != MessageState.DELIVERED_TO_SERVER):
-            return True
-        return self.value < other.value
-
     IN_TRANSIT = 0
     ABORTED = 1
-    CLOSED = 2
-    ERROR = 3
-    DELIVERED_TO_SERVER = 4
-    TIMED_OUT = 5
-    DELIVERED_TO_RECIPIENT = 6
-    SEEN_BY_RECIPIENT = 7
+    ERROR = 2
+    DELIVERED_TO_SERVER = 3
+    DELIVERED_TO_RECIPIENT = 4
+    SEEN_BY_RECIPIENT = 5
 
 
-class MessageTracker(aioxmpp.statemachine.OrderedStateMachine):
+class MessageTracker:
     """
-    This is the high-level equivalent of the :class:`~.StanzaToken`. This
-    structure is used by different tracking implementations.
+    This is the high-level equivalent of the :class:`~.StanzaToken`.
 
-    This is also a :class:`.OrderedStateMachine`, so see there for other
-    methods which allow waiting for a specific state.
+    This structure is used by different tracking implementations. The interface
+    of this class is split in two parts:
 
-    .. attribute:: state
+    1. The public interface for use by applications.
+    2. The "protected" interface for use by tracking implementations.
 
-       The current :class:`MessageState` of the :class:`MessageTracker`. Do
-       **not** write to this attribute from user code. Writing to this
-       attribute is intended only for the tracking implementation.
+    :class:`MessageTracker` objects are designed to be drivable from multiple
+    tracking implementations at once. The idea is that different tracking
+    implementations can cover different parts of the path a stanza takes: one
+    can cover the path to the server (by hooking into the events of a
+    :class:`~.StanzaToken`), the other implementation can use e.g. :xep:`184` to
+    determine delivery at the target and so on.
 
-    .. attribute:: token
+    Methods and attributes from the "protected" interface are marked by a
+    leading underscore.
 
-       The :class:`~.StanzaToken` of the message. This is usually set by the
-       tracking implementation right when the tracker is initialised.
+    .. autoattribute:: state
 
-    .. signal:: on_state_change(state)
+    .. autoattribute:: response
 
-       The signal is emitted with the new state as its only argument when the
-       state of the message tracker changes
+    .. autoattribute:: closed
+
+    .. signal:: on_state_changed(new_state, response=None)
+
+       Emits when a new state is entered.
+
+       :param new_state: The new state of the tracker.
+       :type new_state: :class:`~.MessageState` member
+       :param response: A stanza related to the state.
+       :type response: :class:`~.StanzaBase` or :data:`None`
+
+       The is *not* emitted when the tracker is closed.
+
+    .. signal:: on_closed()
+
+       Emits when the tracker is closed.
+
+    .. automethod:: close
+
+    "Protected" interface:
+
+    .. automethod:: _set_state
 
     """
 
-    on_state_change = aioxmpp.callbacks.Signal()
+    on_closed = aioxmpp.callbacks.Signal()
+    on_state_changed = aioxmpp.callbacks.Signal()
 
-    def __init__(self, token=None):
-        super().__init__(MessageState.IN_TRANSIT)
-        self.token = token
+    def __init__(self):
+        super().__init__()
+        self._state = MessageState.IN_TRANSIT
+        self._response = None
+        self._closed = False
 
-    def on_stanza_state_change(self, token, stanza_state):
-        new_state = self.state
-        if stanza_state == stream.StanzaState.ABORTED:
-            new_state = MessageState.ABORTED
-        elif stanza_state == stream.StanzaState.DROPPED:
-            new_state = MessageState.ABORTED
-        elif stanza_state == stream.StanzaState.ACKED:
-            new_state = MessageState.DELIVERED_TO_SERVER
+    @property
+    def state(self):
+        """
+        The current state of the tracking. Read-only.
+        """
+        return self._state
 
-        if new_state != self.state and not new_state < self.state:
-            self.state = new_state
+    @property
+    def response(self):
+        """
+        A stanza which is relevant to the current state. For
+        :attr:`.MessageState.ERROR`, this will generally be a
+        :class:`.MessageType.ERROR` stanza. For other states, this is either
+        :data:`None` or another stanza depending on the tracking
+        implementation.
+        """
+        return self._response
 
-    @aioxmpp.statemachine.OrderedStateMachine.state.setter
-    def state(self, new_state):
-        aioxmpp.statemachine.OrderedStateMachine.state.fset(self, new_state)
-        self.on_state_change(new_state)
+    @property
+    def closed(self):
+        """
+        Boolean indicator whether the tracker is closed.
+
+        .. seealso::
+
+           :meth:`close` for details.
+        """
+        return self._closed
+
+    def close(self):
+        """
+        Close the tracking, clear all references to the tracker and release all
+        tracking-related resources.
+
+        This operation is idempotent. It does not change the :attr:`state`, but
+        :attr:`closed` turns :data:`True`.
+
+        The :meth:`on_closed` event is only fired on the first call to
+        :meth:`close`.
+        """
+        if self._closed:
+            return
+        self._closed = True
+        self.on_closed()
+
+    # "Protected" Interface
+
+    def _set_state(self, new_state, response=None):
+        """
+        Set the state of the tracker.
+
+        :param new_state: The new state of the tracker.
+        :type new_state: :class:`~.MessageState` member
+        :param response: A stanza related to the new state.
+        :type response: :class:`~.StanzaBase` or :data:`None`
+        :raise ValueError: if a forbidden state transition is attempted.
+        :raise RuntimeError: if the tracker is closed.
+
+        The state of the tracker is set to the `new_state`. The
+        :attr:`response` is also overriden with the new value, no matter if the
+        new or old value is :data:`None` or not. The :meth:`on_state_changed`
+        event is emitted.
+
+        The following transitions are forbidden and attempting to perform them
+        will raise :class:`ValueError`:
+
+        * any state -> :attr:`~.MessageState.IN_TRANSIT`
+        * :attr:`~.MessageState.DELIVERED_TO_RECIPIENT` ->
+          :attr:`~.MessageState.DELIVERED_TO_SERVER`
+        * :attr:`~.MessageState.SEEN_BY_RECIPIENT` ->
+          :attr:`~.MessageState.DELIVERED_TO_RECIPIENT`
+        * :attr:`~.MessageState.SEEN_BY_RECIPIENT` ->
+          :attr:`~.MessageState.DELIVERED_TO_SERVER`
+        * :attr:`~.MessageState.ABORTED` -> any state
+        * :attr:`~.MessageState.ERROR` -> any state
+
+        If the tracker is already :meth:`close`\ -d, :class:`RuntimeError` is
+        raised. This check happens *before* a test is made whether the
+        transition is valid.
+
+        This method is part of the "protected" interface.
+        """
+        if self._closed:
+            raise RuntimeError("message tracker is closed")
+
+        # reject some transitions as documented
+        if     (self._state == MessageState.ABORTED or
+                self._state == MessageState.ERROR or
+                new_state == MessageState.IN_TRANSIT or
+                (self._state == MessageState.DELIVERED_TO_RECIPIENT and
+                 new_state == MessageState.DELIVERED_TO_SERVER) or
+                (self._state == MessageState.SEEN_BY_RECIPIENT and
+                 new_state == MessageState.DELIVERED_TO_SERVER) or
+                (self._state == MessageState.SEEN_BY_RECIPIENT and
+                 new_state == MessageState.DELIVERED_TO_RECIPIENT)):
+            raise ValueError(
+                "message tracker transition from {} to {} not allowed".format(
+                    self._state,
+                    new_state
+                )
+            )
+
+        self._state = new_state
+        self._response = response
+        self.on_state_changed(self._state, self._response)