+++ tests/test_tracking.py
@@ -19,396 +19,241 @@
 # <http://www.gnu.org/licenses/>.
 #
 ########################################################################
+import itertools
 import unittest
 import unittest.mock
 
-from enum import Enum
 
-import aioxmpp.stanza
-import aioxmpp.structs as structs
-import aioxmpp.statemachine
-import aioxmpp.stream as stream
 import aioxmpp.tracking as tracking
 
 
-class TestMessageState(unittest.TestCase):
-    def test_is_enum(self):
-        self.assertTrue(issubclass(
-            tracking.MessageState,
-            Enum
-        ))
-
-    def test_order(self):
-        self.assertLess(
-            tracking.MessageState.IN_TRANSIT,
-            tracking.MessageState.ABORTED
-        )
-
-        self.assertLess(
-            tracking.MessageState.ABORTED,
-            tracking.MessageState.DELIVERED_TO_SERVER,
-        )
-        self.assertLess(
-            tracking.MessageState.ABORTED,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-        )
-        self.assertLess(
-            tracking.MessageState.ABORTED,
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-        )
-
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_SERVER,
-            tracking.MessageState.ABORTED,
-        )
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-            tracking.MessageState.ABORTED,
-        )
-        self.assertLess(
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-            tracking.MessageState.ABORTED,
-        )
-
-        self.assertLess(
-            tracking.MessageState.TIMED_OUT,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-        )
-        self.assertLess(
-            tracking.MessageState.TIMED_OUT,
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-        )
-
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_SERVER,
-            tracking.MessageState.TIMED_OUT,
-        )
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-            tracking.MessageState.TIMED_OUT,
-        )
-        self.assertLess(
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-            tracking.MessageState.TIMED_OUT,
-        )
-
-        self.assertLess(
-            tracking.MessageState.CLOSED,
-            tracking.MessageState.DELIVERED_TO_SERVER,
-        )
-        self.assertLess(
-            tracking.MessageState.CLOSED,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-        )
-        self.assertLess(
-            tracking.MessageState.CLOSED,
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-        )
-
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_SERVER,
-            tracking.MessageState.CLOSED,
-        )
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-            tracking.MessageState.CLOSED,
-        )
-        self.assertLess(
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-            tracking.MessageState.CLOSED,
-        )
-
-        self.assertLess(
-            tracking.MessageState.ERROR,
-            tracking.MessageState.DELIVERED_TO_SERVER,
-        )
-        self.assertLess(
-            tracking.MessageState.ERROR,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-        )
-        self.assertLess(
-            tracking.MessageState.ERROR,
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-        )
-
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_SERVER,
-            tracking.MessageState.ERROR,
-        )
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-            tracking.MessageState.ERROR,
-        )
-        self.assertLess(
-            tracking.MessageState.SEEN_BY_RECIPIENT,
-            tracking.MessageState.ERROR,
-        )
-
-        self.assertLess(
-            tracking.MessageState.TIMED_OUT,
-            tracking.MessageState.ABORTED,
-        )
-        self.assertLess(
-            tracking.MessageState.TIMED_OUT,
-            tracking.MessageState.CLOSED,
-        )
-        self.assertLess(
-            tracking.MessageState.TIMED_OUT,
-            tracking.MessageState.ERROR,
-        )
-
-        self.assertLess(
-            tracking.MessageState.ABORTED,
-            tracking.MessageState.TIMED_OUT,
-        )
-        self.assertLess(
-            tracking.MessageState.ABORTED,
-            tracking.MessageState.CLOSED,
-        )
-        self.assertLess(
-            tracking.MessageState.ABORTED,
-            tracking.MessageState.ERROR,
-        )
-
-        self.assertLess(
-            tracking.MessageState.CLOSED,
-            tracking.MessageState.ABORTED,
-        )
-        self.assertLess(
-            tracking.MessageState.CLOSED,
-            tracking.MessageState.TIMED_OUT,
-        )
-        self.assertLess(
-            tracking.MessageState.CLOSED,
-            tracking.MessageState.ERROR,
-        )
-
-        self.assertFalse(
-            tracking.MessageState.ABORTED <
-            tracking.MessageState.IN_TRANSIT
-        )
-
-        self.assertFalse(
-            tracking.MessageState.ERROR <
-            tracking.MessageState.IN_TRANSIT
-        )
-
-        self.assertFalse(
-            tracking.MessageState.TIMED_OUT <
-            tracking.MessageState.IN_TRANSIT
-        )
-
-        self.assertLess(
-            tracking.MessageState.IN_TRANSIT,
-            tracking.MessageState.DELIVERED_TO_SERVER,
-        )
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_SERVER,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT
-        )
-        self.assertLess(
-            tracking.MessageState.DELIVERED_TO_RECIPIENT,
-            tracking.MessageState.SEEN_BY_RECIPIENT
-        )
-
-
 class TestMessageTracker(unittest.TestCase):
-    def test_is_statemachine(self):
-        self.assertTrue(issubclass(
-            tracking.MessageTracker,
-            aioxmpp.statemachine.OrderedStateMachine
-        ))
-
     def setUp(self):
-        self.stanza = aioxmpp.stanza.Message(
-            type_=structs.MessageType.CHAT
-        )
-        self.token = stream.StanzaToken(self.stanza)
-        self.tr = tracking.MessageTracker(self.token)
-
-    def test_init(self):
-        self.assertIs(self.tr.token, self.token)
-
-    def test_setting_state_emits_on_state_change(self):
-        mock = unittest.mock.Mock()
-        mock.return_value = None
-        self.tr.on_state_change.connect(mock)
-
-        self.tr.state = tracking.MessageState.DELIVERED_TO_SERVER
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_SERVER)
-        mock.assert_called_with(
-            tracking.MessageState.DELIVERED_TO_SERVER
-        )
-
-        self.tr.state = tracking.MessageState.DELIVERED_TO_RECIPIENT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT)
-        mock.assert_called_with(
-            tracking.MessageState.DELIVERED_TO_RECIPIENT
-        )
-
-        self.tr.state = tracking.MessageState.SEEN_BY_RECIPIENT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.SEEN_BY_RECIPIENT)
-        mock.assert_called_with(
-            tracking.MessageState.SEEN_BY_RECIPIENT
-        )
-
-    def test_aborting_after_delivered_to_server_not_possible(self):
-        mock = unittest.mock.Mock()
-        mock.return_value = None
-        self.tr.on_state_change.connect(mock)
-
-        self.tr.state = tracking.MessageState.DELIVERED_TO_SERVER
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_SERVER)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.ABORTED
-
-        self.tr.state = tracking.MessageState.DELIVERED_TO_RECIPIENT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.ABORTED
-
-        self.tr.state = tracking.MessageState.SEEN_BY_RECIPIENT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.SEEN_BY_RECIPIENT)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.ABORTED
-
-        self.assertSequenceEqual(
-            mock.mock_calls,
-            [
-                unittest.mock.call(tracking.MessageState.DELIVERED_TO_SERVER),
-                unittest.mock.call(
-                    tracking.MessageState.DELIVERED_TO_RECIPIENT),
-                unittest.mock.call(tracking.MessageState.SEEN_BY_RECIPIENT),
-            ]
-        )
-
-    def test_timeout_is_possible_after_delivered_to_server(self):
-        mock = unittest.mock.Mock()
-        mock.return_value = None
-        self.tr.on_state_change.connect(mock)
-
-        self.tr.state = tracking.MessageState.DELIVERED_TO_SERVER
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_SERVER)
-
-        self.tr.state = tracking.MessageState.TIMED_OUT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.TIMED_OUT)
-
-    def test_timeout_is_not_possible_after_delivered_to_recipient(self):
-        mock = unittest.mock.Mock()
-        mock.return_value = None
-        self.tr.on_state_change.connect(mock)
-
-        self.tr.state = tracking.MessageState.DELIVERED_TO_RECIPIENT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_RECIPIENT)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.TIMED_OUT
-
-        self.tr.state = tracking.MessageState.SEEN_BY_RECIPIENT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.SEEN_BY_RECIPIENT)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.TIMED_OUT
-
-    def test_delivery_is_not_possible_after_timeout(self):
-        mock = unittest.mock.Mock()
-        mock.return_value = None
-        self.tr.on_state_change.connect(mock)
-
-        self.tr.state = tracking.MessageState.TIMED_OUT
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.TIMED_OUT)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.DELIVERED_TO_RECIPIENT
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.SEEN_BY_RECIPIENT
-
-    def test_delivered_to_server_after_aborting_not_possible(self):
-        self.tr.state = tracking.MessageState.ABORTED
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.ABORTED)
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.DELIVERED_TO_SERVER
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.IN_TRANSIT
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.DELIVERED_TO_RECIPIENT
-
-        with self.assertRaises(ValueError):
-            self.tr.state = tracking.MessageState.SEEN_BY_RECIPIENT
-
-    def test_on_stanza_state_change_to_aborted(self):
-        self.tr.on_stanza_state_change(self.token, stream.StanzaState.ABORTED)
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.ABORTED
-        )
-
-    def test_on_stanza_state_change_to_acked(self):
-        self.tr.on_stanza_state_change(self.token, stream.StanzaState.ACKED)
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_SERVER
-        )
-
-    def test_on_stanza_state_change_to_sent_without_sm(self):
-        self.tr.on_stanza_state_change(self.token, stream.StanzaState.SENT_WITHOUT_SM)
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.IN_TRANSIT
-        )
-
-    def test_on_stanza_state_change_to_sent(self):
-        self.tr.on_stanza_state_change(self.token, stream.StanzaState.SENT)
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.IN_TRANSIT
-        )
-
-    def test_on_stanza_state_change_to_dropped(self):
-        self.tr.on_stanza_state_change(self.token, stream.StanzaState.DROPPED)
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.ABORTED
-        )
-
-    def test_on_stanza_state_change_does_not_fail_on_state_errors(self):
-        self.tr.state = tracking.MessageState.DELIVERED_TO_SERVER
-        self.tr.on_stanza_state_change(self.token, stream.StanzaState.ABORTED)
-        self.assertEqual(
-            self.tr.state,
-            tracking.MessageState.DELIVERED_TO_SERVER
-        )
+        self.t = tracking.MessageTracker()
+        self.listener = unittest.mock.Mock()
+        for ev in ["on_closed", "on_state_changed"]:
+            cb = getattr(self.listener, ev)
+            cb.return_value = None
+            getattr(self.t, ev).connect(cb)
 
     def tearDown(self):
-        del self.tr
+        del self.t
+
+    def test_init(self):
+        self.assertIsNone(self.t.response)
+        self.assertEqual(
+            self.t.state,
+            tracking.MessageState.IN_TRANSIT
+        )
+        self.assertIs(self.t.closed, False)
+
+    def test_state_is_not_writable(self):
+        with self.assertRaises(AttributeError):
+            self.t.state = self.t.state
+
+    def test_response_is_not_writable(self):
+        with self.assertRaises(AttributeError):
+            self.t.response = self.t.response
+
+    def test_closed_is_not_writable(self):
+        with self.assertRaises(AttributeError):
+            self.t.closed = self.t.closed
+
+    def test_close_sets_closed_to_true(self):
+        self.t.close()
+        self.assertIs(self.t.closed, True)
+
+    def test_close_fires_event(self):
+        self.t.close()
+        self.listener.on_closed.assert_called_once_with()
+        self.listener.on_state_changed.assert_not_called()
+
+    def test_close_is_idempotent(self):
+        self.t.close()
+        self.listener.on_closed.assert_called_once_with()
+        self.listener.on_state_changed.assert_not_called()
+        self.t.close()
+        self.listener.on_closed.assert_called_once_with()
+        self.listener.on_state_changed.assert_not_called()
+
+    def test__set_state_updates_state_and_response(self):
+        self.t._set_state(
+            tracking.MessageState.ERROR,
+            unittest.mock.sentinel.response,
+        )
+
+        self.assertEqual(
+            self.t.state,
+            tracking.MessageState.ERROR,
+        )
+        self.assertEqual(
+            self.t.response,
+            unittest.mock.sentinel.response,
+        )
+
+    def test__set_state_fires_event(self):
+        self.t._set_state(
+            tracking.MessageState.DELIVERED_TO_RECIPIENT,
+            unittest.mock.sentinel.response,
+        )
+        self.listener.on_state_changed.assert_called_once_with(
+            tracking.MessageState.DELIVERED_TO_RECIPIENT,
+            unittest.mock.sentinel.response,
+        )
+
+    def test__set_state_rejects_transitions_from_aborted(self):
+        self.t._set_state(
+            tracking.MessageState.ABORTED,
+            unittest.mock.sentinel.response,
+        )
+        self.listener.on_state_changed.reset_mock()
+        self.listener.on_state_changed.return_value = None
+
+        for state in tracking.MessageState:
+            with self.assertRaisesRegex(
+                    ValueError,
+                    "transition from .* to .*not allowed"):
+                self.t._set_state(state)
+            self.listener.on_state_changed.assert_not_called()
+            self.assertEqual(
+                self.t.state,
+                tracking.MessageState.ABORTED,
+            )
+            self.assertEqual(
+                self.t.response,
+                unittest.mock.sentinel.response,
+            )
+
+    def test__set_state_rejects_transitions_from_error(self):
+        self.t._set_state(
+            tracking.MessageState.ERROR,
+            unittest.mock.sentinel.response,
+        )
+        self.listener.on_state_changed.reset_mock()
+        self.listener.on_state_changed.return_value = None
+
+        for state in tracking.MessageState:
+            with self.assertRaisesRegex(
+                    ValueError,
+                    "transition from .* to .*not allowed"):
+                self.t._set_state(state)
+            self.listener.on_state_changed.assert_not_called()
+            self.assertEqual(
+                self.t.state,
+                tracking.MessageState.ERROR,
+            )
+            self.assertEqual(
+                self.t.response,
+                unittest.mock.sentinel.response,
+            )
+
+    def test__set_state_rejects_transitions_to_in_transit(self):
+        for state in tracking.MessageState:
+            t = tracking.MessageTracker()
+            if state != t.state:
+                t._set_state(state, unittest.mock.sentinel.response)
+            on_state_changed = unittest.mock.Mock()
+            t.on_state_changed.connect(on_state_changed)
+            with self.assertRaisesRegex(
+                    ValueError,
+                    "transition from .* to .*not allowed"):
+                t._set_state(tracking.MessageState.IN_TRANSIT)
+            on_state_changed.assert_not_called()
+            self.assertEqual(
+                t.state,
+                state,
+            )
+            if state != t.state:
+                self.assertEqual(
+                    t.response,
+                    unittest.mock.sentinel.response,
+                )
+
+    def test__set_state_rejects_some_other_transitions(self):
+        to_reject = [
+            (
+                tracking.MessageState.DELIVERED_TO_RECIPIENT,
+                tracking.MessageState.DELIVERED_TO_SERVER,
+            ),
+            (
+                tracking.MessageState.SEEN_BY_RECIPIENT,
+                tracking.MessageState.DELIVERED_TO_SERVER,
+            ),
+            (
+                tracking.MessageState.SEEN_BY_RECIPIENT,
+                tracking.MessageState.DELIVERED_TO_RECIPIENT,
+            ),
+        ]
+
+        for state1, state2 in itertools.product(
+                tracking.MessageState,
+                tracking.MessageState):
+            if state1 == tracking.MessageState.ABORTED:
+                # already tested elsewhere
+                continue
+            if state1 == tracking.MessageState.ERROR:
+                # already tested elsewhere
+                continue
+            if state2 == tracking.MessageState.IN_TRANSIT:
+                # already tested elsewhere
+                continue
+            t = tracking.MessageTracker()
+
+            if state1 != t.state:
+                t._set_state(state1, unittest.mock.sentinel.response)
+
+            on_state_changed = unittest.mock.Mock()
+            t.on_state_changed.connect(on_state_changed)
+
+            if (state1, state2) in to_reject:
+                with self.assertRaisesRegex(
+                        ValueError,
+                        "transition from .* to .*not allowed",
+                        msg=(state1, state2)):
+                    t._set_state(state2)
+                on_state_changed.assert_not_called()
+                self.assertEqual(
+                    t.state,
+                    state1,
+                )
+                if state1 != t.state:
+                    self.assertEqual(
+                        t.response,
+                        unittest.mock.sentinel.response,
+                    )
+
+            else:
+                t._set_state(
+                    state2,
+                    unittest.mock.sentinel.response2
+                )
+                self.assertEqual(
+                    t.state,
+                    state2,
+                )
+                self.assertEqual(
+                    t.response,
+                    unittest.mock.sentinel.response2,
+                )
+
+    def test__set_state_bails_out_early_if_closed(self):
+        self.t._set_state(
+            tracking.MessageState.DELIVERED_TO_SERVER,
+            unittest.mock.sentinel.response,
+        )
+
+        self.t.close()
+
+        for s in tracking.MessageState:
+            with self.assertRaisesRegex(
+                    RuntimeError,
+                    "tracker is closed"):
+                self.t._set_state(s)
+            self.assertEqual(
+                self.t.state,
+                tracking.MessageState.DELIVERED_TO_SERVER,
+            )
+            self.assertEqual(
+                self.t.response,
+                unittest.mock.sentinel.response,
+            )